Add Jest setup and tests
