+++ README.md
@@ -1,0 +1,7 @@
+# BubbleLog
+
+This project uses Jest for unit testing. After cloning the repository and installing dependencies with `npm install`, run tests with:
+
+```bash
+npm test
+```
